gmail_sender/requirements.txt
PyYAML>=5.4.1
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6 